DT-835 RenameConvert conviction controller tests to unit tests
